# Optimize API V2 Links endpoint and CustomerMailer: eliminate N+1 queries 