--- a/app/controllers/api/v2/links_controller.rb
+++ b/app/controllers/api/v2/links_controller.rb
@@ -9,13 +9,6 @@ class Api::V2::LinksController < Api::V2::BaseController
 
   def index
     products = current_resource_owner.products.visible
-      .select(
-        :id, :user_id, :name, :description, :require_shipping, :preview_url,
-        :custom_receipt, :customizable_price, :custom_permalink, :subscription_duration,
-        :price_currency_type, :deleted_at, :max_purchase_count, :purchase_disabled_at,
-        :banned_at, :flags, :price_cents, :rental_price_cents,
-        :unique_permalink, :native_type, :purchase_type, :json_data, :created_at, :taxonomy_id
-      )
       .includes(
         :tags,
         :asset_previews,
@@ -26,7 +19,6 @@ class Api::V2::LinksController < Api::V2::BaseController
         :custom_fields,
         :taxonomy,
         { user: [:bank_accounts, :merchant_accounts, :custom_fields] },
-        { prices: :link },
         variant_categories_alive: { alive_variants: :variant_category }
       ).order(created_at: :desc)