--- a/app/models/link.rb
+++ b/app/models/link.rb
@@ -1352,6 +1352,10 @@ class Link < ApplicationRecord
 
       ppp_factors = purchasing_power_parity_enabled? ? options[:preloaded_ppp_factors] || PurchasingPowerParityService.new.get_all_countries_factors(user) : nil
 
+      preloaded_tags = tags.loaded? ? tags : tags.includes(:product_taggings)
+      preloaded_prices = prices.loaded? ? prices.alive.is_buy : prices.alive.is_buy.includes(:link)
+      preloaded_variant_categories = variant_categories_alive.loaded? ? variant_categories_alive : variant_categories_alive.includes(alive_variants: [:variant_category])
+
       json = super_as_json(only: keep).merge!(
         "id" => external_id,
         "url" => nil, # Deprecated
@@ -1359,7 +1363,7 @@ class Link < ApplicationRecord
         "currency" => price_currency_type,
         "short_url" => long_url,
         "thumbnail_url" => thumbnail&.alive&.url.presence,
-        "tags" => tags.pluck(:name),
+        "tags" => preloaded_tags.map(&:name),
         "formatted_price" => price_formatted_verbose,
         "published" => alive?,
         "file_info" => multifile_aware_product_file_info,
@@ -1368,8 +1372,8 @@ class Link < ApplicationRecord
         "custom_fields" => custom_field_descriptors.as_json,
         "custom_summary" => custom_summary,
         "is_tiered_membership" => is_tiered_membership?,
-        "recurrences" => is_tiered_membership? ? prices.alive.is_buy.map(&:recurrence).uniq : nil,
-        "variants" => variant_categories_alive.map do |cat|
+        "recurrences" => is_tiered_membership? ? preloaded_prices.map(&:recurrence).uniq : nil,
+        "variants" => preloaded_variant_categories.map do |cat|
           {
             title: cat.title,
             options: cat.alive_variants.map do |variant|