@sankalp1999 2025-06-28
```
# Quick one-liners for Rails console testing

# Simple query counter
def sql_count(&block)
  count = 0
  ActiveSupport::Notifications.subscribe("sql.active_record") { count += 1 }
  result = block.call
  ActiveSupport::Notifications.unsubscribe("sql.active_record")
  puts "#{count} SQL queries executed"
  result
end

# Test your specific API endpoint optimization
def test_api_optimization(user_id = nil)
  user = user_id ? User.find(user_id) : User.first

  puts "=== BEFORE optimization ==="
  sql_count do
    # Old way (simulate removing the new includes)
    products = user.products.visible.includes(:preorder_link, :tags, :taxonomy, variant_categories_alive: [:alive_variants]).limit(10)
    products.map { |p| p.as_json(api_scopes: ["view_sales"]) }
  end

  puts "\n=== AFTER optimization ==="
  sql_count do
    # New way with all the includes
    products = user.products.visible.includes(
      :tags,
      :asset_previews,
      :thumbnail,
      :preorder_link,
      :alive_prices,
      :product_files,
      :custom_fields,
      :taxonomy,
      { user: [:bank_accounts, :merchant_accounts, :custom_fields] },
      { prices: :link },
      variant_categories_alive: { alive_variants: :variant_category }
    ).limit(10)
    products.map { |p| p.as_json(api_scopes: ["view_sales"]) }
  end
end

# Test the URL redirect presenter optimization
def test_url_redirect_optimization
  purchase = Purchase.joins(:link).first

  puts "=== URL Redirect Presenter Test ==="
  puts "Testing with purchase ID: #{purchase.id}"

  # You can call this to test the before/after of your url_redirect_presenter.rb changes
  sql_count do
    # This will use your optimized version
    UrlRedirectPresenter.new(purchase).present_post_purchase_content
  end
end

# Test customer mailer optimization
def test_customer_mailer_optimization
  purchase_ids = Purchase.limit(3).pluck(:id)

  puts "=== Customer Mailer Test ==="

  puts "BEFORE (without includes):"
  sql_count do
    Purchase.where(id: purchase_ids).map { |p| Charge::Chargeable.find_by_purchase_or_charge!(purchase: p) }
  end

  puts "AFTER (with includes):"
  sql_count do
    purchases = Purchase.where(id: purchase_ids).includes(:charge).to_a
    purchases.map { |p| Charge::Chargeable.find_by_purchase_or_charge!(purchase: p) }
  end
end

# Usage in Rails console:
# test_api_optimization
# test_customer_mailer_optimization
# test_url_redirect_optimization
```