@sankalp1999 2025-07-09
ok i fixed this. i did some investigation and i actually missed a field so i am removing select. can be problematic if it breaks stuff.

https://github.com/antiwork/gumroad/pull/488#issuecomment-3052239262